Extract GettingStarted into sub-components
